packages/nextjs/requirements.txt
numpy 
Flask 
stable-baselines3
gymnasium 
plotly 
pandas 
requests 
yfinance 
matplotlib 
scikit-learn 
pytz 
python-dotenv 
APScheduler 
diskcache 
flipside
